TODO: debug

+++ find_best_match.h
@@ -1,3 +1,3 @@
 #include "image.h"
 
-double find_best_match(Image *Ga, Image *Gs, int L, int xs, int ys);
+int find_best_match(Image *Ga, Image *Gs, int L, int xs, int ys);
